First pass at adding encode support
